@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <UserSecretsId>aspnet-Dersa_C-E14BE20D-7DE9-4E97-AE2F-6B7438B1F54A</UserSecretsId>
   </PropertyGroup>
 
@@ -16,6 +16,7 @@
     <Compile Remove="Controllers\ConfigController.cs" />
     <Compile Remove="Controllers\ProjectController1.cs" />
     <Compile Remove="Controllers\Startup.Auth.cs" />
+    <Compile Remove="IQueryExecuteService.cs" />
     <Compile Remove="Models\AccountViewModels.cs" />
   </ItemGroup>
 
@@ -29,34 +30,42 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.1.2" PrivateAssets="All" />
+    <PackageReference Include="System.Data.SqlClient" Version="4.8.1" />
   </ItemGroup>
 
   <ItemGroup>
+    <Reference Include="DbManager">
+      <HintPath>bin\Debug\netcoreapp2.1\DbManager.dll</HintPath>
+    </Reference>
     <Reference Include="Dersa.Common">
       <HintPath>..\Dersa\bin\Dersa.Common.dll</HintPath>
     </Reference>
     <Reference Include="Dersa.Interfaces">
       <HintPath>..\Dersa\bin\Dersa.Interfaces.dll</HintPath>
     </Reference>
-    <Reference Include="DIOS.BusinessBase">
-      <HintPath>..\..\gitd\dersa\DIOS\DIOS.Server\common\DIOS.BusinessBase.dll</HintPath>
+    <Reference Include="Dersa.Stereotypes">
+      <HintPath>..\Dersa\bin\Dersa.Stereotypes.dll</HintPath>
     </Reference>
     <Reference Include="Dios.Common">
-      <HintPath>..\..\gitd\dersa\DIOS\bin\Dios.Common.dll</HintPath>
+      <HintPath>bin\Debug\netcoreapp2.1\Dios.Common.dll</HintPath>
     </Reference>
     <Reference Include="Dios.Interfaces">
-      <HintPath>..\..\gitd\dersa\DIOS\DIOS.Server\common\Dios.Interfaces.dll</HintPath>
+      <HintPath>bin\Debug\netcoreapp2.1\Dios.Interfaces.dll</HintPath>
     </Reference>
     <Reference Include="Dios.SqlManager">
-      <HintPath>..\..\gitd\dersa\DIOS\DIOS.Server\common\Dios.SqlManager.dll</HintPath>
+      <HintPath>bin\Debug\netcoreapp2.1\Dios.SqlManager.dll</HintPath>
+    </Reference>
+    <Reference Include="DIOS.WCF.Core">
+      <HintPath>bin\Debug\netcoreapp2.1\DIOS.WCF.Core.dll</HintPath>
     </Reference>
     <Reference Include="Microsoft.AspNet.Identity.Core">
-      <HintPath>..\..\VSP\Web\OwinServer\KatanaApp\bin\Debug\Microsoft.AspNet.Identity.Core.dll</HintPath>
+      <HintPath>bin\Debug\netcoreapp2.1\Microsoft.AspNet.Identity.Core.dll</HintPath>
     </Reference>
     <Reference Include="Microsoft.Owin.Security">
-      <HintPath>..\..\VSP\Web\OwinServer\KatanaApp\bin\Debug\Microsoft.Owin.Security.dll</HintPath>
+      <HintPath>bin\Debug\netcoreapp2.1\Microsoft.Owin.Security.dll</HintPath>
+    </Reference>
+    <Reference Include="Newtonsoft.Json">
+      <HintPath>..\..\gitd\dersa\DIOS\DIOS.Server\common\Newtonsoft.Json.dll</HintPath>
     </Reference>
     <Reference Include="System.Web.Optimization">
       <HintPath>..\Dersa\bin\System.Web.Optimization.dll</HintPath>
