@@ -42,13 +42,6 @@
             width: 100%;
         }
     </style>
-    <script src="dist/vue.js"></script>
-    <script src="dist/jquery.js"></script>
-    <script src="dist/jquery-ui.js"></script>
-    <script src="dist/jstree.js"></script>
-    <script src="dist/dersatree.js"></script>
-    <script src="js/mxClient.js"></script>
-
     <script type="text/javascript">
         var mxBasePath = '../../';
     </script>
@@ -86,16 +79,7 @@
     </style>
 </head>
 <body onload="createEditor('config/diagrameditor.xml');">
-    <!--
-    <button id="find_button">Найти</button>
-    <input id="searchtext" type="text" size="105" /> <br>
-    <input id="found" type="text" size="10" value="0" />
-    <input id="foundpath" type="text" size="80" />
-    <button id="goto_button">Перейти</button>
-    <button id="reset_button">Reset</button><br>
-    <select id="listBox" multiple size="5"></select>
-    <div style="margin: 0 0 0 auto"> Имя пользователя: <strong><span style="color: green;">@ViewBag.Login</span></strong></div>
-        -->
+
     <div style="float:left;width:200px;margin-top:-20px"><h1>DERSA modeler</h1></div>
     <div style="position:absolute;width: 260px;top: 96px; left: 300px;"><input id="tbSearch" style="width:100%" type="text"></div>
     <div style="position:absolute;width:80px;top:92px;left:556px">
@@ -180,17 +164,6 @@
             form.window = wnd;
             wnd.setVisible(true);
         });
-        //var n = 5;
-        //function on_load_node(node, status) {
-        //    console.log(node);
-        //    if (node.children.length && n > 0) {
-        //        $('#dersa').jstree(true).load_node(node.children[0], on_load_node);
-        //        //n--;
-        //    }
-        //    else {
-        //        $('#dersa').jstree(true).select_node('602250');
-        //    }
-        //}
 
     </script>
     <div id="infoboard" class="block2">
@@ -207,6 +180,13 @@
             }
         });
     </script>
+    <script src="dist/vue.js"></script>
+    <script src="dist/jquery.js"></script>
+    <script src="dist/jquery-ui.js"></script>
+    <script src="dist/jstree.js"></script>
+    <script src="dist/dersatree.js"></script>
+    <script src="js/mxClient.js"></script>
+
     @*@section Scripts {
             @Scripts.Render("~/bundles/jstree")
             @Scripts.Render("~/dist/dersatree.js")
@@ -515,46 +495,6 @@
 
             editor.addAction('initG', initGraph);
 
-            // Create select actions in page
-            //var node = document.getElementById('mainActions');
-            //var buttons = ['group', 'ungroup', 'cut', 'copy', 'paste', 'delete', 'undo', 'redo', 'print', 'show'];
-
-
-            //for (var i = 0; i < buttons.length; i++) {
-            //    var button = document.createElement('button');
-            //    mxUtils.write(button, mxResources.get(buttons[i]));
-
-            //    var factory = function (name) {
-            //        return function () {
-            //            editor.execute(name);
-            //        };
-            //    };
-
-            //    mxEvent.addListener(button, 'click', factory(buttons[i]));
-            //    node.appendChild(button);
-            //}
-
-            //// Create select actions in page
-            //var node = document.getElementById('selectActions');
-            //mxUtils.write(node, 'Select: ');
-            //mxUtils.linkAction(node, 'All', editor, 'selectAll');
-            //mxUtils.write(node, ', ');
-            //mxUtils.linkAction(node, 'None', editor, 'selectNone');
-            //mxUtils.write(node, ', ');
-            //mxUtils.linkAction(node, 'Vertices', editor, 'selectVertices');
-            //mxUtils.write(node, ', ');
-            //mxUtils.linkAction(node, 'Edges', editor, 'selectEdges');
-
-            //// Create select actions in page
-            //var node = document.getElementById('zoomActions');
-            //mxUtils.write(node, 'Zoom: ');
-            //mxUtils.linkAction(node, 'In', editor, 'zoomIn');
-            //mxUtils.write(node, ', ');
-            //mxUtils.linkAction(node, 'Out', editor, 'zoomOut');
-            //mxUtils.write(node, ', ');
-            //mxUtils.linkAction(node, 'Actual', editor, 'actualSize');
-            //mxUtils.write(node, ', ');
-            //mxUtils.linkAction(node, 'Fit', editor, 'fit');
         }
 
                                             //window.onbeforeunload = function () { return mxResources.get('changesLost'); };
@@ -576,9 +516,9 @@
                     <!-- Toolbar Here -->
                 </td>
                 <td valign="top" style="border-width:1px;border-style:solid;border-color:black;">
-                    <div id="graph" tabindex="-1" style="height:480px;width:684px;display:none;overflow:hidden;cursor:default;">
+                    <div id="graph" tabindex="-1" style="height:480px;width:800px;display:none;overflow:hidden;cursor:default;">
                     </div>
-                    <div id="diagram" tabindex="-1" style="height:800px;width:684px;display:none;overflow:hidden;cursor:default;">
+                    <div id="diagram" tabindex="-1" style="height:800px;width:800px;display:none;overflow:hidden;cursor:default;">
                     </div>
                     <textarea id="xml" style="height:480px;width:684px;display:none;border-style:none;"></textarea>
                 </td>
